- cleanups - changed x64 code changes in texture stuff to same as linux port
Imported-from: https://subversion.assembla.com/svn/AvP/branches/workbranch/trunk@283

+++ 3dc/avp/support/scstring.hpp
@@ -181,7 +181,7 @@
 		return bCanRender[ I_Font ];
 	}
 
-	inline unsigned int SCString::GetNumChars(void)
+	inline size_t SCString::GetNumChars(void)
 	{
 		return NumberOfCharacters;
 	}
